+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
  - All references can now be obtained from the `ReferenceMap` via `listReferences()` (#73)
 
+### Fixed
+ - Fix 0-based ordered lists starting at 1 instead of 0 (#74)
+
 ## [0.7.0] - 2015-02-16
 ### Added
  - More unit tests to increase code coverage